Start work on editing registered time

+++ src/main/java/schedulingapp/Activity.java
@@ -56,7 +56,7 @@ public class Activity {
 		
 	}
 	
-	public void editTime(Developer dev, int hours, Calendar date) {
+	public void editTime(Developer dev, double hours, Calendar date) {
 		
 	}
